refactor(ui): split usage chart renderers
